Used legend for descript and fig legend for shotnums.  Cleaned up code
It turns out that only one legend can be associated with a plot axes at
a given time which means that if I used two plt.legend command, only the
second one will be displayed.  The work-around is to use plt.gca() to
add the legend to the axes again.

It turns out that plt.figlegend() has no such restriction and when I went
from using plt.legend() to plt.figlegend(), the I made the plt.gca()
command obsolete.

The code has been cleaned up to reflect this new neatness.

For right now, the descript will be a legend for the plot whereas the
figlegend will be associated with the figure of the plot.