Metadata files to speed things up
I created metadata files for the imacon timing and for the date that the
shot was taken.

With the imacon timing, I will be able to extract the timing of the plasma
without having to explicitly read the imacon files themselves.

I will no longer need to call IDL to extract a foldername for singleloop shots.