Nonfunctional attempt to use python to edit IDL files
Python doesn't write to binary in the same way that IDL does.
I attempted to correct the magnetic data and redo it in a way that
is compatible with both IDL and Python.

But this isn't a good use of my time.  So I'm going to just make reduced
hall data and store that instead.